requirements.txt
Flask==3.1.0
requests==2.32.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
langchain==0.2.0
langchain-community==0.2.0
gunicorn==22.0.0
